Truck reversing sound: removed React.Profiler
React Profiler's console.log output was interfering with Ink's terminal rendering in development mode, causing the "Cwd:" line to disappear during navigation.
To be honest, then we didn't get any valuable information from the console.log since we do not have a dev console like we do in a browser; and we open up the app in an alternate buffer screen that doesn't display the log. We'll just rely on the React devtool instead.