fix: Empty Git branches would cause infinite "Loading branches..." state
By "Empty Git branches" I mean: if you spin up Git Rake in a new project that only has the `main` brach, and we exclude the main branch, then `branches.length === 0` and we relied on the amount of branches in several places.
This update introduced a generic "isLoading" state and fix the issue so we correctly show the "No branches found" message instead.