feat: Handle restore and trashing branch name conflicts
The GitRepository.ts code wasn't written in a way that allows for recursive calls to Git batch operations - at least not in a way that would make the code easy to maintain down the road.
So, after accepting that the code had code-smell, a lot of the code have been made drastically simpler and DRY'er so we can trashing/restoring in `trashBranchesWithRetry` and `restoreBranchesWithRetry` functions.