refactor: Migrate away from Git notes for branch deletion dates
Using `Git notes` initially sounded like a good idea since it's a native Git solution meant for storing meta-data, but as I've started doing some performance improvmenets, I realised that we called `git notes add ...` for each branch we move to trash which leads to O(N) and we'd optimally like to do O(1).
I then started looking into bulk-creating notes with `update-ref` and it turns out `git notes add` does quite a bit of work under the hood; e.g. each note is designed to point to a commit tree and if we create a blob and point to that SHA instead of a tree, our metadata isn't showing up when using `git notes list` or `git notes show`, so we're "breaking" the original design intent of Git.
Well, alternatively, we could also use `git fast-import` to create trees etc., but the whole solutions starts to become overly complex for the value we get.

Instead, we've refactored trash branches to suffix a `@2025-07-29` (YYYY-MM-DD) at the end of trashed branches; we're stripping the date as part of the new creating a GitTrashBranch so `branch.name` can still be used cleanly in the UI (backwards compatible) and `branch.ref` still contains the full ref to be used for actions.
As a small bonus, then we no longer have to keep two different refs in sync (think: deleting and creating a note each time we do something with a trash branch) which reduces some of the edge-cases we have to handle.

More context if you're curious: we do not want to store a branch's deleted date on a commit SHA since it can be shared by multiple branches. User experience: you have two branches (branch-1 & branch-2). If you trashed branch-1 months ago and you trashed branch-2 yesterday, the cleanup script will remove both of them because they share the same deletion date. Technically it should be OK to delete both since the "SHA" doesn't matter but the user experience is that a branch you just trashed completely "disappeared" instead of being in the trash-bin until TTL is reached.