feat: Merge confirmation mode into Branches list + Selected filter
Before this change, the "confirmation mode" (basically the Y/n prompt to trash, delete, restore)  would create a separate experience where the selected branches couldn't be navigated and our ScrollableList component would have to do additional "available height calculations".
After this change, the confirmation prompt lives inside the StatusBar at the top and it's now possible to navigate selected branches and deselect if anything appears by mistake.

Being able to navigate the selected branches revealed some issue in the codebase that I fixed in the same go to get it working:
The "current list of branches to work with" was being derived and consumed/manipulted from ~3 different places (filtered list, available list, selected list).
We now have 1 useBranchesState() hook that handles all of that logic, and BranchDataContext + SelectionContext has been merged into a single context (BranchesSelectionContext) to further simplify how we work with branches.
This should mean that we only have a single source of truth for working with selected and available branches.

Bonus rewards from this:
We're adding a new "Selected" filter type. This is both avaialble through the normal filter cycle keybinding, but it also gives us a fitler we can programmatically toggle in "confirmation mode" so we can navigate <BranchesList> in a built-in way and avoid duplicate maintenance in the future.