feat(branches list): Move navigation index up or down after selection
This is inspired by fzf where each selection moves the user to a new item.
For some reason (just my personal instinct) we use the space key for selections, and not tab like fzf, but it works the same when regular space moves the index down and shift+space moves the index upwards.