fix(deps): Roll back React + Ink versions
When we use React 19 / Ink 6, we're seeing issues in our alternate buffer screen where all content is moved up 1 line and we're missing a line at the bottom.
This is a temporary downgrade but Vadim (Ink creator) have said that Ink wasn't created to fullscreen TUI (https://github.com/vadimdemedes/ink/issues/182#issuecomment-483095617). So I also consider this an open invitation to consider if we should:

1. Investigate why the new React 19 update is messing with alternate buffer screens
2. Look into the complications of migrating to something like https://github.com/infely/react-curse